pyproject.toml
[tool.poetry]
name = "repl_python3_keys"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
replit = "^3.2.4"
aiopyston = "^1.1.1"
Quart = "^0.15.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
